投稿者以外が削除できないよう修正

+++ app/src/Controller/AnswersController.php
@@ -60,7 +60,14 @@ class AnswersController extends AppController
   public function delete(int $id)
   {
     $answer = $this->Answers->get($id);
+    $questionId = $answer->question_id;
+
     // 回答を削除できるのは回答投稿者のみとする
+    if ($answer->user_id !== $this->Auth->user('id')) {
+      $this->Flash->error('他のユーザーの回答を削除することはできません');
+
+      return $this->redirect(['controller' => 'Questions', 'action' => 'view', $questionId]);
+    }
 
     if ($this->Answers->delete($answer)) {
       $this->Flash->success('回答を削除しました');
